backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bardapi==1.0.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
browser-cookie3==0.19.1
capstone==5.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
colored-traceback==0.3.0
contextlib2==21.6.0
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-auto-logout==0.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-ninja==1.1.0
django-ninja-extra==0.20.5
django-ninja-jwt==5.3.0
django-ranged-response==0.2.0
django-restframework==0.0.1
django-simple-captcha==0.6.0
django-soet==1.0
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
inflection==0.5.1
injector==0.21.0
intervaltree==3.1.0
loguru==0.7.2
lz4==4.3.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
paramiko==3.4.0
pillow==10.3.0
plumbum==1.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
puremagic==1.21
pwntools==4.12.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pyelftools==0.31
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-gemini-api==2.4.2
python3-openid==3.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
ROPGadget==7.4
rpyc==6.0.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12
typing_extensions==4.11.0
tzdata==2024.1
unicorn==2.0.1.post1
unix-ar==0.2.1
uritemplate==4.1.1
urllib3==2.2.1
vigenere==1.1.0
wcwidth==0.2.13
wikipedia==1.4.0
win32-setctime==1.1.0
WMI==1.5.1
xortool==1.0.2
zipp==3.18.1
zstandard==0.22.0